pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dso_system"
version = "0.2.4"
description = "Distributed scraping infrastructure with Prefect and Playwright."
readme = "README.md"
requires-python = ">=3.13" # 3.14が登場しても動くように設定
license = { text = "Proprietary" }

# メインの依存関係
dependencies = [
    "requests",
    "beautifulsoup4",
    "playwright",
    "prefect>=3.0,<4.0",
    "psycopg2-binary",
    "pyyaml",
    "pandas",        # CSV/Excel処理用
    "boto3>=1.34.0", # AWS S3 アップロード用
]

# パッケージ発見設定:
# `uv pip install -e .` を一度実行すると、プロジェクトルートが Python パスに永続登録される。
# 以降は `from src.framework.static import ...` がどのファイル・どのIDEから実行しても動作する。
# sys.path のコードを各クローラーファイルに書く必要がなくなる。
[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]
exclude = ["scripts*", "tests*", "docs*", "DevKit*"]
